fix: add suspense
